Use IHostedService for DLNA
